Separate instead of Env annotations, improve docs
